Move compile_stage to compile.py
