+++ pippy/PipelineStage.py
@@ -1,16 +1,13 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
 import logging
-from typing import Dict, List, Tuple
 
 import torch
 import torch.distributed as dist
-from torch._subclasses.fake_tensor import FakeTensorMode
 import pippy
 
 from pippy.fx.passes import shape_prop
-from pippy.IR import MultiUseParameterConfig, Pipe
+from pippy.IR import Pipe
 from pippy.microbatch import merge_chunks, split_args_kwargs_into_chunks
-from pippy.utils import PIPPY_VERBOSITY
 
 
 def _make_tensor_from_meta(
@@ -163,58 +160,3 @@ class PipelineStage(torch.nn.Module):
             None, #self.output_chunk_spec,
             False, #self._debug_mask_minibatches
         )
-
-
-def compile_stage(
-    mod: torch.nn.Module,
-    rank: int,
-    num_ranks: int,
-    num_chunks: int,
-    device: torch.device,
-    group: dist.ProcessGroup,
-    example_inputs: List[torch.Tensor],
-) -> PipelineStage:
-    # Trace and cut
-    pipe = Pipe.from_tracing(mod, MultiUseParameterConfig.REPLICATE)
-    gm = pipe.split_gm
-    if rank == 0:
-        logging.info(gm)
-        if PIPPY_VERBOSITY == "INFO":
-            gm.graph.print_tabular()
-
-    args_split, _ = split_args_kwargs_into_chunks(
-        example_inputs,
-        {},  # kwargs included in `example_inputs`
-        num_chunks,
-        None, #self.args_chunk_spec,
-        None, #self.kwargs_chunk_spec,
-        False, #self._debug_mask_minibatches,
-    )
-
-    # Use fake tensor for shape propagation
-    # Since model itself may have been materialized, we need to use
-    # `allow_non_fake_inputs`
-    fake_mode = FakeTensorMode(allow_non_fake_inputs=True)
-    # In reality, the fake input should be created from shape info (potentially
-    # broadcast from Rank 0)
-    fake_args_split = pippy.fx.node.map_aggregate(
-        args_split,
-        lambda a: fake_mode.from_tensor(a)
-    )
-
-    # Use 1st chunk of args for shape propagation
-    chunk0 = fake_args_split[0]
-
-    sp = shape_prop.ShapeProp(gm)
-    sp.propagate(*chunk0)
-
-    # Create pipeline stage
-    return PipelineStage(
-        pipe,
-        rank,
-        num_ranks,
-        num_chunks,
-        device,
-        group,
-    )
-
